Add Auto Zeus!
This makes it possible to become Zeus on any mission. Makes it possible
for us to save a bad mission from ruining the entire evening; smaller
bugs can be fixed right there in the mission, or we could rebalance
missions that might end up too easy or too hard. We can also play
missions built for a larger player group by having a Zeus HQ that can
grant airstrikes, etc.