Unit locator: Tweak ranges, add "Way off"
You will now see, in the editor, how precise the different values are. I
hope this helps you select appropriate values for your missions.

I increased the inaccuracy of all values slightly because I felt that
the precision was still a bit too great in all current offerings. The
"Way off" marker would be best used with "Which town is populated right
now?" rather than "Which building is populated?". A convoy might be a
good fit as well.